MAGETWO-84756: Prepare codebase for 2.0.18

+++ app/code/Magento/Cms/Controller/Adminhtml/Wysiwyg/Images/DeleteFolder.php
@@ -6,6 +6,9 @@
  */
 namespace Magento\Cms\Controller\Adminhtml\Wysiwyg\Images;
 
+use \Magento\Framework\App\ObjectManager;
+use \Magento\Framework\App\Filesystem\DirectoryList;
+
 class DeleteFolder extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images
 {
     /**
@@ -19,19 +22,29 @@ class DeleteFolder extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images
     protected $resultRawFactory;
 
     /**
+     * @var \Magento\Framework\App\Filesystem\DirectoryResolver
+     */
+    private $directoryResolver;
+
+    /**
      * @param \Magento\Backend\App\Action\Context $context
      * @param \Magento\Framework\Registry $coreRegistry
      * @param \Magento\Framework\Controller\Result\JsonFactory $resultJsonFactory
      * @param \Magento\Framework\Controller\Result\RawFactory $resultRawFactory
+     * @param \Magento\Framework\App\Filesystem\DirectoryResolver|null $directoryResolver
+     * @throws \RuntimeException
      */
     public function __construct(
         \Magento\Backend\App\Action\Context $context,
         \Magento\Framework\Registry $coreRegistry,
         \Magento\Framework\Controller\Result\JsonFactory $resultJsonFactory,
-        \Magento\Framework\Controller\Result\RawFactory $resultRawFactory
+        \Magento\Framework\Controller\Result\RawFactory $resultRawFactory,
+        \Magento\Framework\App\Filesystem\DirectoryResolver $directoryResolver = null
     ) {
         $this->resultRawFactory = $resultRawFactory;
         $this->resultJsonFactory = $resultJsonFactory;
+        $this->directoryResolver = $directoryResolver
+            ?: ObjectManager::getInstance()->get(\Magento\Framework\App\Filesystem\DirectoryResolver::class);
         parent::__construct($context, $coreRegistry);
     }
 
@@ -44,6 +57,11 @@ class DeleteFolder extends \Magento\Cms\Controller\Adminhtml\Wysiwyg\Images
     {
         try {
             $path = $this->getStorage()->getCmsWysiwygImages()->getCurrentPath();
+            if (!$this->directoryResolver->validatePath($path, DirectoryList::MEDIA)) {
+                throw new \Magento\Framework\Exception\LocalizedException(
+                    __('Directory %1 is not under storage root path.', $path)
+                );
+            }
             $this->getStorage()->deleteDirectory($path);
             return $this->resultRawFactory->create();
         } catch (\Exception $e) {
